+++ aws-sdk-core/lib/aws-sdk-core/gamelift.rb
@@ -1,0 +1,5 @@
+Aws.add_service(:GameLift, {
+  api: "#{Aws::API_DIR}/gamelift/2015-10-01/api-2.json",
+  docs: "#{Aws::API_DIR}/gamelift/2015-10-01/docs-2.json",
+  examples: "#{Aws::API_DIR}/gamelift/2015-10-01/examples-1.json",
+})